+++ .vscode/tasks.json
@@ -6,20 +6,7 @@
         "type": "shell",
         "command": "npm run build -w @equal-vote/star-vote-shared",
       },
-      {
-        "label": "Port-Forward PostgreSQL",
-        "type": "shell",
-        "command": "kubectl port-forward service/postgresql -n starvote 5432:5432",
-        "isBackground": true,
-        "problemMatcher": {
-          "base": "$tsc-watch",
-          "background": {
-            "activeOnStart": true,
-            "beginsPattern": ".",
-            "endsPattern": "Forwarding from"
-          }
-        }
-      },
+      // await page.goto('http://localhost:3000/');
       {
         "label": "Start Frontend",
         "type": "shell",
