+++ docker-compose.yml
@@ -6,17 +6,17 @@ services:
   # we should probably pull from the repository so it doesn't take forever to build the container.
     build: . 
     ports:
-      - "5000:5000"
+      - "5000:3000"
     networks:
       - star-net
     environment:
         - REACT_APP_FEATURED_ELECTIONS=''
         - REACT_APP_MAX_BALLOT_RANKS=10
         - REACT_APP_DEFAULT_BALLOT_RANKS=6
-        - REACT_APP_BACKEND_URL=http://localhost:5000
+        - REACT_APP_BACKEND_URL=http://localhost:3000
         - REACT_APP_KEYCLOAK_URL=http://localhost:8080/realms/Dev/protocol/openid-connect
         - DANGEROUSLY_DISABLE_HOST_CHECK=true
-        - FRONTEND_PORT=5000
+        - FRONTEND_PORT=3000
         - DATABASE_URL=postgresql://postgres:ChangeMeOrDontTest2020@my-db:5432/postgres
         - DEV_DATABASE=FALSE
         - KEYCLOAK_SECRET=DefaultKeycloakSecret
