+++ testing/playwright/auth/user.json
@@ -1,0 +1,143 @@
+{
+  "cookies": [
+    {
+      "name": "AUTH_SESSION_ID",
+      "value": "81b70037-07d8-47ea-b5f8-04dd99436504",
+      "domain": "localhost",
+      "path": "/realms/Dev/",
+      "expires": -1,
+      "httpOnly": true,
+      "secure": true,
+      "sameSite": "None"
+    },
+    {
+      "name": "AUTH_SESSION_ID_LEGACY",
+      "value": "81b70037-07d8-47ea-b5f8-04dd99436504",
+      "domain": "localhost",
+      "path": "/realms/Dev/",
+      "expires": -1,
+      "httpOnly": true,
+      "secure": false,
+      "sameSite": "Lax"
+    },
+    {
+      "name": "KEYCLOAK_IDENTITY",
+      "value": "eyJhbGciOiJIUzUxMiIsInR5cCIgOiAiSldUIiwia2lkIiA6ICJhMWFmMjFiNy03ODEwLTRiMjQtYWNjNC02MzFlODA2MWI3NWQifQ.eyJleHAiOjE3NDUyOTg2MDgsImlhdCI6MTc0NDQzNDYwOCwianRpIjoiMDIwZWQ3MTktZWQ4ZC00NzQxLTlkNDItODYyNDU3ODlkOWYyIiwiaXNzIjoiaHR0cDovL2xvY2FsaG9zdDo4MDgwL3JlYWxtcy9EZXYiLCJzdWIiOiI2ZWIwODRjNy1jNTcyLTQwZDEtYmY1My0yNmE2MjA5NTI4NWUiLCJ0eXAiOiJTZXJpYWxpemVkLUlEIiwic2Vzc2lvbl9zdGF0ZSI6IjgxYjcwMDM3LTA3ZDgtNDdlYS1iNWY4LTA0ZGQ5OTQzNjUwNCIsInNpZCI6IjgxYjcwMDM3LTA3ZDgtNDdlYS1iNWY4LTA0ZGQ5OTQzNjUwNCIsInN0YXRlX2NoZWNrZXIiOiJWeFRZVG9DeHJlZi1vaE9adVg4SzB1LXBmenduTFVnbkYwajgzcXdydVZVIn0.bCtOJsRh4ZcJoy-zyDi-qiU8u7F9uiypGWaSR-Vy1BiVj-QnxHX8FCNsNFZHUF65wDZn1QX_Whko7pJPJ8bLoQ",
+      "domain": "localhost",
+      "path": "/realms/Dev/",
+      "expires": -1,
+      "httpOnly": true,
+      "secure": true,
+      "sameSite": "None"
+    },
+    {
+      "name": "KEYCLOAK_IDENTITY_LEGACY",
+      "value": "eyJhbGciOiJIUzUxMiIsInR5cCIgOiAiSldUIiwia2lkIiA6ICJhMWFmMjFiNy03ODEwLTRiMjQtYWNjNC02MzFlODA2MWI3NWQifQ.eyJleHAiOjE3NDUyOTg2MDgsImlhdCI6MTc0NDQzNDYwOCwianRpIjoiMDIwZWQ3MTktZWQ4ZC00NzQxLTlkNDItODYyNDU3ODlkOWYyIiwiaXNzIjoiaHR0cDovL2xvY2FsaG9zdDo4MDgwL3JlYWxtcy9EZXYiLCJzdWIiOiI2ZWIwODRjNy1jNTcyLTQwZDEtYmY1My0yNmE2MjA5NTI4NWUiLCJ0eXAiOiJTZXJpYWxpemVkLUlEIiwic2Vzc2lvbl9zdGF0ZSI6IjgxYjcwMDM3LTA3ZDgtNDdlYS1iNWY4LTA0ZGQ5OTQzNjUwNCIsInNpZCI6IjgxYjcwMDM3LTA3ZDgtNDdlYS1iNWY4LTA0ZGQ5OTQzNjUwNCIsInN0YXRlX2NoZWNrZXIiOiJWeFRZVG9DeHJlZi1vaE9adVg4SzB1LXBmenduTFVnbkYwajgzcXdydVZVIn0.bCtOJsRh4ZcJoy-zyDi-qiU8u7F9uiypGWaSR-Vy1BiVj-QnxHX8FCNsNFZHUF65wDZn1QX_Whko7pJPJ8bLoQ",
+      "domain": "localhost",
+      "path": "/realms/Dev/",
+      "expires": -1,
+      "httpOnly": true,
+      "secure": false,
+      "sameSite": "Lax"
+    },
+    {
+      "name": "KEYCLOAK_SESSION",
+      "value": "\"Dev/6eb084c7-c572-40d1-bf53-26a62095285e/81b70037-07d8-47ea-b5f8-04dd99436504\"",
+      "domain": "localhost",
+      "path": "/realms/Dev/",
+      "expires": 1745298608.292455,
+      "httpOnly": false,
+      "secure": true,
+      "sameSite": "None"
+    },
+    {
+      "name": "KEYCLOAK_SESSION_LEGACY",
+      "value": "\"Dev/6eb084c7-c572-40d1-bf53-26a62095285e/81b70037-07d8-47ea-b5f8-04dd99436504\"",
+      "domain": "localhost",
+      "path": "/realms/Dev/",
+      "expires": 1745298608.292484,
+      "httpOnly": false,
+      "secure": false,
+      "sameSite": "Lax"
+    },
+    {
+      "name": "_pk_id.2.1fff",
+      "value": "ac22a45c90bc8278.1744434607.",
+      "domain": "localhost",
+      "path": "/",
+      "expires": 1778389807,
+      "httpOnly": false,
+      "secure": false,
+      "sameSite": "Lax"
+    },
+    {
+      "name": "_pk_ses.2.1fff",
+      "value": "1",
+      "domain": "localhost",
+      "path": "/",
+      "expires": 1744436408,
+      "httpOnly": false,
+      "secure": false,
+      "sameSite": "Lax"
+    },
+    {
+      "name": "temp_id",
+      "value": "aa33b64d-3f51-49bb-adc0-217dd6c48d67",
+      "domain": "localhost",
+      "path": "/",
+      "expires": -1,
+      "httpOnly": false,
+      "secure": false,
+      "sameSite": "Lax"
+    },
+    {
+      "name": "access_token",
+      "value": "eyJhbGciOiJSUzI1NiIsInR5cCIgOiAiSldUIiwia2lkIiA6ICJtalZBR2l4ZS1ZblpqSV9RbEJKVzZGMkM4dF9SbWRqREVNZExDV3k4LUMwIn0.eyJleHAiOjE3NDQ4NjY2MDgsImlhdCI6MTc0NDQzNDYwOCwiYXV0aF90aW1lIjoxNzQ0NDM0NjA4LCJqdGkiOiI4ODEwZjlmMi1kOTRiLTRhYjQtOTMzNC1mMjMwZGI3ODhlMzkiLCJpc3MiOiJodHRwOi8vbG9jYWxob3N0OjgwODAvcmVhbG1zL0RldiIsImF1ZCI6ImFjY291bnQiLCJzdWIiOiI2ZWIwODRjNy1jNTcyLTQwZDEtYmY1My0yNmE2MjA5NTI4NWUiLCJ0eXAiOiJCZWFyZXIiLCJhenAiOiJ3ZWIiLCJzZXNzaW9uX3N0YXRlIjoiODFiNzAwMzctMDdkOC00N2VhLWI1ZjgtMDRkZDk5NDM2NTA0IiwiYWNyIjoiMSIsInJlYWxtX2FjY2VzcyI6eyJyb2xlcyI6WyJkZWZhdWx0LXJvbGVzLXN0YXIgdm90aW5nIGRldiIsIm9mZmxpbmVfYWNjZXNzIiwidW1hX2F1dGhvcml6YXRpb24iXX0sInJlc291cmNlX2FjY2VzcyI6eyJhY2NvdW50Ijp7InJvbGVzIjpbIm1hbmFnZS1hY2NvdW50IiwibWFuYWdlLWFjY291bnQtbGlua3MiLCJ2aWV3LXByb2ZpbGUiXX19LCJzY29wZSI6Im9wZW5pZCBlbWFpbCBwcm9maWxlIiwic2lkIjoiODFiNzAwMzctMDdkOC00N2VhLWI1ZjgtMDRkZDk5NDM2NTA0IiwiZW1haWxfdmVyaWZpZWQiOmZhbHNlLCJuYW1lIjoiVGVzdCBUZXN0IiwicHJlZmVycmVkX3VzZXJuYW1lIjoicGxheXdyaWdodHRlc3QiLCJnaXZlbl9uYW1lIjoiVGVzdCIsImZhbWlseV9uYW1lIjoiVGVzdCIsImVtYWlsIjoidGVzdEB0ZXN0LmNvbSJ9.AqNvVLnCO_XamRrG6CCKHLLnChZgecq1m5MQekniAasKB1fmshSJqwlMAXiiHCoK7tObc5LamqJ7GrRSKTSapxouuXgP6Pw-mYYnmxrBSNiZyeg0zuN4X0k-PXMGiuIRGILs_SaVwhSxoZay3yErki6Ecejco1N0t5eDAFaeO2LPkHNe72XJvRWLJrqe6fs7YQS1QPT6vElWgG7QxdXuYYm1yrCecft9GSmSSGq1ZKLzb03pJZ_869MFrBzJ0DsaVdglD_LV-T21xspOuya997GoqqO64YXxzRrmVxxct103N_Ab2fETGiYSbmHPuMJ-yAieUvB1JKp29I3TsdEyWA",
+      "domain": "localhost",
+      "path": "/",
+      "expires": 1744866609,
+      "httpOnly": false,
+      "secure": false,
+      "sameSite": "Lax"
+    },
+    {
+      "name": "id_token",
+      "value": "eyJhbGciOiJSUzI1NiIsInR5cCIgOiAiSldUIiwia2lkIiA6ICJtalZBR2l4ZS1ZblpqSV9RbEJKVzZGMkM4dF9SbWRqREVNZExDV3k4LUMwIn0.eyJleHAiOjE3NDQ4NjY2MDgsImlhdCI6MTc0NDQzNDYwOCwiYXV0aF90aW1lIjoxNzQ0NDM0NjA4LCJqdGkiOiJhZWZmMDcwZS0wZWIyLTQ0MzUtOGEwZS0yZDg3YThlNjA1NGUiLCJpc3MiOiJodHRwOi8vbG9jYWxob3N0OjgwODAvcmVhbG1zL0RldiIsImF1ZCI6IndlYiIsInN1YiI6IjZlYjA4NGM3LWM1NzItNDBkMS1iZjUzLTI2YTYyMDk1Mjg1ZSIsInR5cCI6IklEIiwiYXpwIjoid2ViIiwic2Vzc2lvbl9zdGF0ZSI6IjgxYjcwMDM3LTA3ZDgtNDdlYS1iNWY4LTA0ZGQ5OTQzNjUwNCIsImF0X2hhc2giOiJ3c18zUzBtQ0tFb1o4bTd6MmhtdTJRIiwiYWNyIjoiMSIsInNpZCI6IjgxYjcwMDM3LTA3ZDgtNDdlYS1iNWY4LTA0ZGQ5OTQzNjUwNCIsImVtYWlsX3ZlcmlmaWVkIjpmYWxzZSwibmFtZSI6IlRlc3QgVGVzdCIsInByZWZlcnJlZF91c2VybmFtZSI6InBsYXl3cmlnaHR0ZXN0IiwiZ2l2ZW5fbmFtZSI6IlRlc3QiLCJmYW1pbHlfbmFtZSI6IlRlc3QiLCJlbWFpbCI6InRlc3RAdGVzdC5jb20ifQ.cOWZX9aR2Vd_L0TYKaMB8Bj-R3yZE42i8IuHA2e89nSR64xUgujUQNYkZkttRM31jBOcoCVEg25QGWix-gDX6wNoq4Zgapj1PSlyvV9ToMhduMc-cgBQThzltunt-YkIHAN5JEKYOvnhTt175buifKZ2uHOyJLBh1vZlG0e8lgBfgIfYVUIWI4j7aBuCl4NfVegGQQpx35_MJKaRXqbVFpf9ldXjmvsnoAWIh8SmxMi6amlOAQQyD5XXvI6Pz_qMhv6skfWXebbpP-Coh4VwBRO847SYlcrpxSNGAOYlxf3hdq9CcH0H2XFKigFfbpLEXVHbfv8YwIvYSNY5EN3aMQ",
+      "domain": "localhost",
+      "path": "/",
+      "expires": 1744866609,
+      "httpOnly": false,
+      "secure": false,
+      "sameSite": "Lax"
+    },
+    {
+      "name": "refresh_token",
+      "value": "eyJhbGciOiJIUzUxMiIsInR5cCIgOiAiSldUIiwia2lkIiA6ICJhMWFmMjFiNy03ODEwLTRiMjQtYWNjNC02MzFlODA2MWI3NWQifQ.eyJleHAiOjE3NDQ0MzY0MDgsImlhdCI6MTc0NDQzNDYwOCwianRpIjoiNWJhMTcyNTUtYzMzMS00YWE1LTg4NTgtMGU3ZGU3NmQ2OTUzIiwiaXNzIjoiaHR0cDovL2xvY2FsaG9zdDo4MDgwL3JlYWxtcy9EZXYiLCJhdWQiOiJodHRwOi8vbG9jYWxob3N0OjgwODAvcmVhbG1zL0RldiIsInN1YiI6IjZlYjA4NGM3LWM1NzItNDBkMS1iZjUzLTI2YTYyMDk1Mjg1ZSIsInR5cCI6IlJlZnJlc2giLCJhenAiOiJ3ZWIiLCJzZXNzaW9uX3N0YXRlIjoiODFiNzAwMzctMDdkOC00N2VhLWI1ZjgtMDRkZDk5NDM2NTA0Iiwic2NvcGUiOiJvcGVuaWQgZW1haWwgcHJvZmlsZSIsInNpZCI6IjgxYjcwMDM3LTA3ZDgtNDdlYS1iNWY4LTA0ZGQ5OTQzNjUwNCJ9.bEOyWlVFB9_L7kWTOhktpFzSYLRf-gSzSK6l9Xru8zqd2gHM2n9UiSWgVeMI2UAeaaDolLau1P2rdN9iYy_ZwA",
+      "domain": "localhost",
+      "path": "/",
+      "expires": 1744866609,
+      "httpOnly": false,
+      "secure": false,
+      "sameSite": "Lax"
+    }
+  ],
+  "origins": [
+    {
+      "origin": "http://localhost:3000",
+      "localStorage": [
+        {
+          "name": "i18nextLng",
+          "value": "en-US"
+        },
+        {
+          "name": "themeMode",
+          "value": "\"browserDefault\""
+        },
+        {
+          "name": "flag_overrides",
+          "value": "{}"
+        }
+      ]
+    }
+  ]
+}
