+++ testing/tests/test-1.spec.ts
@@ -1,5 +1,8 @@
 import { test, expect } from '@playwright/test';
 
 test('test', async ({ page }) => {
-  // Recording...
+  await page.goto('http://localhost:3000/');
+  await page.getByRole('textbox', { name: 'What is your poll question?' }).click();
+  await page.getByRole('textbox', { name: 'What is your poll question?' }).fill('I hate this');
 });
+
